legend for clips stats
