--- web/src/components/pages/new-home/clip-stats.css
@@ -1,58 +1,0 @@
-@import url('../../vars.css');
-
-.home-card {
-    width: 100%;
-    max-width: 570px;
-    padding: 30px;
-    box-sizing: border-box;
-    background: var(--white);
-    box-shadow: var(--card-shadow);
-
-    & svg {
-        height: 165px;
-    }
-
-    & .tick-label {
-        font-size: 12px;
-        fill: var(--warm-grey);
-    }
-
-    & path {
-        &.total {
-            stroke: var(--valid-green);
-        }
-        &.valid {
-            stroke: var(--red);
-        }
-        &.unverified {
-            stroke: var(--blue);
-        }
-    }
-
-    & circle {
-        &.inner {
-            &.total {
-                fill: var(--valid-green);
-            }
-            &.valid {
-                fill: var(--red);
-            }
-            &.unverified {
-                fill: var(--blue);
-            }
-        }
-
-        &.outer {
-            &.total {
-                filter: drop-shadow(0 3px 3px var(--valid-green));
-            }
-            &.valid {
-                filter: drop-shadow(0 3px 3px var(--red));
-            }
-            &.unverified {
-                filter: drop-shadow(0 3px 3px var(--blue));
-            }
-        }
-    }
-
-}
