+++ web/src/components/pages/new-home/home.css
@@ -75,6 +75,7 @@
 
         & .show-more {
             border: none;
+            padding: 0;
             font-size: 12px;
             font-weight: 600;
             letter-spacing: 0.4px;
